worker/requirements.txt

psycopg2-binary
redis
pika
